+++ packages/kofono/src/builder/types.ts
@@ -1,5 +1,4 @@
 import type { Form } from "../form/Form";
-import type { SchemaPlugin } from "../plugins/types";
 import type { Property } from "../property/Property";
 import type { Schema, SchemaProperty } from "../schema/Schema";
 import type { Token } from "../schema/Tokens";
@@ -8,10 +7,10 @@ import type { PropertyDeclaration } from "./k/PropertyDeclaration";
 export interface PropertyBuilder<TSchemaType extends SchemaProperty> {
     buildProperty(): Property<TSchemaType>;
 }
-
-export interface PluginBuilder<TPlugin extends SchemaPlugin> {
-    buildPlugin(): Promise<TPlugin>;
-}
+//
+// export interface PluginBuilder<TPlugin extends SchemaExtension> {
+//     buildPlugin(): Promise<TPlugin>;
+// }
 
 export type SchemaBuildResult =
     | SchemaBuildResultSucceed
@@ -30,20 +29,20 @@ export type SchemaBuildResultFailed = {
 export interface SchemaPropertiesDeclarations {
     $id?: Schema[Token.SchemaId];
     $vars?: Schema[Token.SchemaVars];
-    $plugins?: Schema[Token.SchemaPlugins];
+    $extensions?: Schema[Token.SchemaExtensions];
     $translations?: Schema[Token.SchemaTranslations];
     [
         key: Exclude<
             string,
             | Token.SchemaId
             | Token.SchemaVars
-            | Token.SchemaPlugins
+            | Token.SchemaExtensions
             | Token.SchemaTranslations
         >
     ]:
         | PropertyDeclaration
         | Schema[Token.SchemaId]
         | Schema[Token.SchemaVars]
-        | Schema[Token.SchemaPlugins]
+        | Schema[Token.SchemaExtensions]
         | Schema[Token.SchemaTranslations];
 }