+++ packages/kofono/src/extension/UpdateCounter/UpdateCounterExtension.ts
@@ -1,35 +1,36 @@
 import { Events } from "../../form/events/types";
 import type { Form } from "../../form/Form";
-import { BasePlugin } from "../BasePlugin";
+import { BaseExtension } from "../BaseExtension";
 
-// this is a basic plugin to count the number of updates performed on the form
-// it serves as an example of how to create a custom plugin
+// this is a basic extension to count the number of updates performed on the form
+// it serves as an example of how to create a custom extension
 
 // represent the schema structure definition
-export type SchemaUpdateCounterPlugin = {
+export type SchemaUpdateCounterExtension = {
     updateCounter: UpdateCounterOpts;
 };
 
-// represent the options passed to the plugin at form creation
+// represent the options passed to the extension at form creation
 export type UpdateCounterOpts = unknown;
 
-// represent the plugin meta-state shape/type
+// represent the extension meta-state shape/type
 export type UpdateCounterMeta = number;
 
-// factory to build the plugin instance from the schema definition
-export const updateCounterPluginFactory = {
-    updateCounter: (opts: UpdateCounterOpts) => new UpdateCounterPlugin(opts),
+// factory to build the extension instance from the schema definition
+export const updateCounterExtensionFactory = {
+    updateCounter: (opts: UpdateCounterOpts) =>
+        new UpdateCounterExtension(opts),
 };
 
 // function to build the schema definition
-export function updateCounter(): SchemaUpdateCounterPlugin {
+export function updateCounter(): SchemaUpdateCounterExtension {
     return {
         updateCounter: {},
     };
 }
 
-// the plugin implementation
-export class UpdateCounterPlugin extends BasePlugin<UpdateCounterMeta> {
+// the extension implementation
+export class UpdateCounterExtension extends BaseExtension<UpdateCounterMeta> {
     public readonly name: string = "updateCounter";
     public readonly version: string = "1.0.0";
     public readonly defaultMeta = 0;