+++ packages/kofono/src/extension/defaultExtensionsFactory.ts
@@ -1,0 +1,11 @@
+import type { Extension } from "./types";
+import {
+    updateCounterExtensionFactory
+} from "./UpdateCounter/UpdateCounterExtension";
+
+export const defaultExtensionsFactory: Record<
+    string,
+    (opts: any) => Extension
+> = {
+    ...updateCounterExtensionFactory,
+};