+++ packages/kofono/src/plugins/defaultPluginsFactory.ts
@@ -1,6 +1,0 @@
-import type { Plugin } from "./types";
-import { updateCounterPluginFactory } from "./UpdateCounter/UpdateCounterPlugin";
-
-export const defaultPluginsFactory: Record<string, (opts: any) => Plugin> = {
-    ...updateCounterPluginFactory,
-};